Bevel demo is being implemented
The bevel demo is not able to show a part of the face after the change being applied. The apply change has yet to be implemented. And this demo shall also create a loop cut to make life easier for the designers by avoiding ruining any face-loops which they have created